@@ -5,6 +5,8 @@ package execution
 
 import (
 	"context"
+	"errors"
+	"fmt"
 	"net"
 	"net/http"
 	"sync"
@@ -13,7 +15,6 @@ import (
 
 	"github.com/offchainlabs/nitro/util/redisutil"
 	"github.com/offchainlabs/nitro/util/stopwaiter"
-	"github.com/pkg/errors"
 	flag "github.com/spf13/pflag"
 
 	"github.com/ethereum/go-ethereum/arbitrum"
@@ -78,7 +79,7 @@ func AddOptionsForForwarderConfigImpl(prefix string, defaultConfig *ForwarderCon
 }
 
 type TxForwarder struct {
-	enabled   int32
+	enabled   atomic.Bool
 	target    string
 	timeout   time.Duration
 	transport *http.Transport
@@ -129,7 +130,7 @@ func (f *TxForwarder) ctxWithTimeout(inctx context.Context) (context.Context, co
 }
 
 func (f *TxForwarder) PublishTransaction(inctx context.Context, tx *types.Transaction, options *arbitrum_types.ConditionalOptions) error {
-	if atomic.LoadInt32(&f.enabled) == 0 {
+	if !f.enabled.Load() {
 		return ErrNoSequencer
 	}
 	ctx, cancelFunc := f.ctxWithTimeout(inctx)
@@ -144,7 +145,7 @@ const cacheUpstreamHealth = 2 * time.Second
 const maxHealthTimeout = 10 * time.Second
 
 func (f *TxForwarder) CheckHealth(inctx context.Context) error {
-	if atomic.LoadInt32(&f.enabled) == 0 {
+	if !f.enabled.Load() {
 		return ErrNoSequencer
 	}
 	f.healthMutex.Lock()
@@ -166,7 +167,7 @@ func (f *TxForwarder) Initialize(inctx context.Context) error {
 	if f.target == "" {
 		f.rpcClient = nil
 		f.ethClient = nil
-		f.enabled = 0
+		f.enabled.Store(false)
 		return nil
 	}
 	ctx, cancelFunc := f.ctxWithTimeout(inctx)
@@ -177,13 +178,13 @@ func (f *TxForwarder) Initialize(inctx context.Context) error {
 	}
 	f.rpcClient = rpcClient
 	f.ethClient = ethclient.NewClient(rpcClient)
-	f.enabled = 1
+	f.enabled.Store(true)
 	return nil
 }
 
 // Disable is not thread-safe vs. Initialize
 func (f *TxForwarder) Disable() {
-	atomic.StoreInt32(&f.enabled, 0)
+	f.enabled.Store(false)
 }
 
 func (f *TxForwarder) Start(ctx context.Context) error {
@@ -268,7 +269,7 @@ func (f *RedisTxForwarder) Initialize(ctx context.Context) error {
 	var err error
 	f.redisCoordinator, err = redisutil.NewRedisCoordinator(f.config.RedisUrl)
 	if err != nil {
-		return errors.Wrap(err, "unable to create redis coordinator")
+		return fmt.Errorf("unable to create redis coordinator: %w", err)
 	}
 	f.update(ctx)
 	return nil
@@ -369,7 +370,7 @@ func (f *RedisTxForwarder) Start(ctx context.Context) error {
 		return err
 	}
 	if err := f.CallIterativelySafe(f.update); err != nil {
-		return errors.Wrap(err, "failed to start forwarder update thread")
+		return fmt.Errorf("failed to start forwarder update thread: %w", err)
 	}
 	return nil
 }
