@@ -5,6 +5,7 @@ package execution
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"math"
 	"math/big"
@@ -14,6 +15,7 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/offchainlabs/nitro/arbutil"
 	"github.com/offchainlabs/nitro/util/arbmath"
 	"github.com/offchainlabs/nitro/util/containers"
 	"github.com/offchainlabs/nitro/util/headerreader"
@@ -34,7 +36,6 @@ import (
 	"github.com/offchainlabs/nitro/arbos/arbostypes"
 	"github.com/offchainlabs/nitro/arbos/l1pricing"
 	"github.com/offchainlabs/nitro/util/stopwaiter"
-	"github.com/pkg/errors"
 )
 
 var (
@@ -184,9 +185,8 @@ func (c *nonceCache) matches(header *types.Header) bool {
 		// The header is updated as the block is built,
 		// so instead of checking its hash, we do a pointer comparison.
 		return c.dirty == header
-	} else {
-		return c.block == header.ParentHash
 	}
+	return c.block == header.ParentHash
 }
 
 func (c *nonceCache) Reset(block common.Hash) {
@@ -633,7 +633,7 @@ func (s *Sequencer) expireNonceFailures() *time.Timer {
 // There's no guarantee that returned tx nonces will be correct
 func (s *Sequencer) precheckNonces(queueItems []txQueueItem) []txQueueItem {
 	bc := s.execEngine.bc
-	latestHeader := bc.CurrentBlock().Header()
+	latestHeader := bc.CurrentBlock()
 	latestState, err := bc.StateAt(latestHeader.Root)
 	if err != nil {
 		log.Error("failed to get current state to pre-check nonces", "err", err)
@@ -927,12 +927,14 @@ func (s *Sequencer) createBlock(ctx context.Context) (returnValue bool) {
 	return madeBlock
 }
 
-func (s *Sequencer) updateLatestL1Block(header *types.Header) {
+func (s *Sequencer) updateLatestParentChainBlock(header *types.Header) {
 	s.L1BlockAndTimeMutex.Lock()
 	defer s.L1BlockAndTimeMutex.Unlock()
-	if s.l1BlockNumber < header.Number.Uint64() {
-		s.l1BlockNumber = header.Number.Uint64()
+
+	l1BlockNumber := arbutil.ParentHeaderToL1BlockNumber(header)
+	if header.Time > s.l1Timestamp || (header.Time == s.l1Timestamp && l1BlockNumber > s.l1BlockNumber) {
 		s.l1Timestamp = header.Time
+		s.l1BlockNumber = l1BlockNumber
 	}
 }
 
@@ -945,7 +947,7 @@ func (s *Sequencer) Initialize(ctx context.Context) error {
 	if err != nil {
 		return err
 	}
-	s.updateLatestL1Block(header)
+	s.updateLatestParentChainBlock(header)
 	return nil
 }
 
@@ -967,7 +969,7 @@ func (s *Sequencer) Start(ctxIn context.Context) error {
 					if !ok {
 						return
 					}
-					s.updateLatestL1Block(header)
+					s.updateLatestParentChainBlock(header)
 				case <-ctx.Done():
 					return
 				}
@@ -982,10 +984,9 @@ func (s *Sequencer) Start(ctxIn context.Context) error {
 		if madeBlock {
 			// Note: this may return a negative duration, but timers are fine with that (they treat negative durations as 0).
 			return time.Until(nextBlock)
-		} else {
-			// If we didn't make a block, try again immediately.
-			return 0
 		}
+		// If we didn't make a block, try again immediately.
+		return 0
 	})
 
 	return nil
