@@ -11,15 +11,18 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 )
 
-func CorrespondingL1BlockNumber(ctx context.Context, client L1Interface, blockNumber uint64) (uint64, error) {
-	header, err := client.HeaderByNumber(ctx, big.NewInt(int64(blockNumber)))
-	if err != nil {
-		return 0, fmt.Errorf("error getting L1 block number %d header : %w", blockNumber, err)
-	}
+func ParentHeaderToL1BlockNumber(header *types.Header) uint64 {
 	headerInfo := types.DeserializeHeaderExtraInformation(header)
-	if headerInfo.L1BlockNumber != 0 {
-		return headerInfo.L1BlockNumber, nil
-	} else {
-		return blockNumber, nil
+	if headerInfo.ArbOSFormatVersion > 0 {
+		return headerInfo.L1BlockNumber
 	}
+	return header.Number.Uint64()
+}
+
+func CorrespondingL1BlockNumber(ctx context.Context, client L1Interface, parentBlockNumber uint64) (uint64, error) {
+	header, err := client.HeaderByNumber(ctx, big.NewInt(int64(parentBlockNumber)))
+	if err != nil {
+		return 0, fmt.Errorf("error getting L1 block number %d header : %w", parentBlockNumber, err)
+	}
+	return ParentHeaderToL1BlockNumber(header), nil
 }
