@@ -8,14 +8,13 @@ import (
 	"fmt"
 
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/pkg/errors"
 )
 
 // GetLogEmitterTxData requires that the tx's data is at least 4 bytes long
 func GetLogEmitterTxData(ctx context.Context, client L1Interface, log types.Log) ([]byte, error) {
 	tx, err := client.TransactionInBlock(ctx, log.BlockHash, log.TxIndex)
 	if err != nil {
-		return nil, errors.WithStack(err)
+		return nil, err
 	}
 	if tx.Hash() != log.TxHash {
 		return nil, fmt.Errorf("L1 client returned unexpected transaction hash %v when looking up block %v transaction %v with expected hash %v", tx.Hash(), log.BlockHash, log.TxIndex, log.TxHash)
