@@ -41,9 +41,8 @@ func GetChainConfig(chainId *big.Int, chainName string, genesisBlockNum uint64, 
 	}
 	if chainId.Uint64() != 0 {
 		return nil, fmt.Errorf("missing chain config for L2 chain ID %v", chainId)
-	} else {
-		return nil, fmt.Errorf("missing chain config for L2 chain name %v", chainName)
 	}
+	return nil, fmt.Errorf("missing chain config for L2 chain name %v", chainName)
 }
 
 func GetRollupAddressesConfig(chainId uint64, chainName string, l2ChainInfoFiles []string, l2ChainInfoJson string) (RollupAddresses, error) {
@@ -56,9 +55,8 @@ func GetRollupAddressesConfig(chainId uint64, chainName string, l2ChainInfoFiles
 	}
 	if chainId != 0 {
 		return RollupAddresses{}, fmt.Errorf("missing rollup addresses for L2 chain ID %v", chainId)
-	} else {
-		return RollupAddresses{}, fmt.Errorf("missing rollup addresses for L2 chain name %v", chainName)
 	}
+	return RollupAddresses{}, fmt.Errorf("missing rollup addresses for L2 chain name %v", chainName)
 }
 
 func ProcessChainInfo(chainId uint64, chainName string, l2ChainInfoFiles []string, l2ChainInfoJson string) (*ChainInfo, error) {
@@ -85,9 +83,8 @@ func ProcessChainInfo(chainId uint64, chainName string, l2ChainInfoFiles []strin
 	}
 	if chainId != 0 {
 		return nil, fmt.Errorf("unsupported chain ID %v", chainId)
-	} else {
-		return nil, fmt.Errorf("unsupported chain name %v", chainName)
 	}
+	return nil, fmt.Errorf("unsupported chain name %v", chainName)
 }
 
 func findChainInfo(chainId uint64, chainName string, chainsInfoBytes []byte) (*ChainInfo, error) {
