@@ -138,9 +138,8 @@ func (dbs *DBStorageService) Close(ctx context.Context) error {
 func (dbs *DBStorageService) ExpirationPolicy(ctx context.Context) (arbstate.ExpirationPolicy, error) {
 	if dbs.discardAfterTimeout {
 		return arbstate.DiscardAfterDataTimeout, nil
-	} else {
-		return arbstate.KeepForever, nil
 	}
+	return arbstate.KeepForever, nil
 }
 
 func (dbs *DBStorageService) String() string {
