@@ -33,7 +33,6 @@ type hash = common.Hash
 type bytes32 = [32]byte
 type ctx = *precompiles.Context
 
-type Message = types.Message
 type BackendAPI = core.NodeInterfaceBackendAPI
 type ExecutionResult = core.ExecutionResult
 
@@ -49,18 +48,18 @@ func init() {
 	_, nodeInterfaceDebug := precompiles.MakePrecompile(nodeInterfaceDebugMeta, nodeInterfaceDebugImpl)
 
 	core.InterceptRPCMessage = func(
-		msg Message,
+		msg *core.Message,
 		ctx context.Context,
 		statedb *state.StateDB,
 		header *types.Header,
 		backend core.NodeInterfaceBackendAPI,
-	) (Message, *ExecutionResult, error) {
-		to := msg.To()
+	) (*core.Message, *ExecutionResult, error) {
+		to := msg.To
 		arbosVersion := arbosState.ArbOSVersion(statedb) // check ArbOS has been installed
 		if to != nil && arbosVersion != 0 {
 			var precompile precompiles.ArbosPrecompile
 			var swapMessages bool
-			returnMessage := &Message{}
+			returnMessage := &core.Message{}
 			var address addr
 
 			switch *to {
@@ -99,16 +98,16 @@ func init() {
 			core.ReadyEVMForL2(evm, msg)
 
 			output, gasLeft, err := precompile.Call(
-				msg.Data(), address, address, msg.From(), msg.Value(), false, msg.Gas(), evm,
+				msg.Data, address, address, msg.From, msg.Value, false, msg.GasLimit, evm,
 			)
 			if err != nil {
 				return msg, nil, err
 			}
 			if swapMessages {
-				return *returnMessage, nil, nil
+				return returnMessage, nil, nil
 			}
 			res := &ExecutionResult{
-				UsedGas:       msg.Gas() - gasLeft,
+				UsedGas:       msg.GasLimit - gasLeft,
 				Err:           nil,
 				ReturnData:    output,
 				ScheduledTxes: nil,
@@ -118,7 +117,7 @@ func init() {
 		return msg, nil, nil
 	}
 
-	core.InterceptRPCGasCap = func(gascap *uint64, msg Message, header *types.Header, statedb *state.StateDB) {
+	core.InterceptRPCGasCap = func(gascap *uint64, msg *core.Message, header *types.Header, statedb *state.StateDB) {
 		if *gascap == 0 {
 			// It's already unlimited
 			return
@@ -139,7 +138,7 @@ func init() {
 		}
 
 		posterCost, _ := state.L1PricingState().PosterDataCost(msg, l1pricing.BatchPosterAddress)
-		posterCostInL2Gas := arbos.GetPosterGas(state, header.BaseFee, msg.RunMode(), posterCost)
+		posterCostInL2Gas := arbos.GetPosterGas(state, header.BaseFee, msg.TxRunMode, posterCost)
 		*gascap = arbmath.SaturatingUAdd(*gascap, posterCostInL2Gas)
 	}
 
