@@ -5,6 +5,7 @@ package staker
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"math/big"
 	"time"
@@ -20,7 +21,6 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/offchainlabs/nitro/arbutil"
 	"github.com/offchainlabs/nitro/solgen/go/rollupgen"
-	"github.com/pkg/errors"
 )
 
 type ConfirmType uint8
@@ -286,14 +286,13 @@ func (v *L1Validator) generateNodeAction(ctx context.Context, stakerInfo *OurSta
 			log.Error("invalid start global state inbox position", startState.GlobalState.BlockHash, "batch", startState.GlobalState.Batch, "pos", startState.GlobalState.PosInBatch)
 			return nil, false, errors.New("invalid start global state inbox position")
 		}
-		latestHeader := v.l2Blockchain.CurrentBlock().Header()
+		latestHeader := v.l2Blockchain.CurrentBlock()
 		if latestHeader.Number.Int64() < expectedBlockHeight {
 			log.Info("catching up to chain blocks", "localBlocks", latestHeader.Number, "target", expectedBlockHeight)
 			return nil, false, nil
-		} else {
-			log.Error("unknown start block hash", "hash", startState.GlobalState.BlockHash, "batch", startState.GlobalState.Batch, "pos", startState.GlobalState.PosInBatch)
-			return nil, false, errors.New("unknown start block hash")
 		}
+		log.Error("unknown start block hash", "hash", startState.GlobalState.BlockHash, "batch", startState.GlobalState.Batch, "pos", startState.GlobalState.PosInBatch)
+		return nil, false, errors.New("unknown start block hash")
 	}
 
 	var lastBlockValidated uint64
@@ -319,7 +318,7 @@ func (v *L1Validator) generateNodeAction(ctx context.Context, stakerInfo *OurSta
 			return nil, false, fmt.Errorf("wasmroot doesn't match rollup : %v, valid: %v", v.lastWasmModuleRoot, validRoots)
 		}
 	} else {
-		lastBlockValidated = v.l2Blockchain.CurrentBlock().Header().Number.Uint64()
+		lastBlockValidated = v.l2Blockchain.CurrentBlock().Number.Uint64()
 
 		if localBatchCount > 0 {
 			messageCount, err := v.inboxTracker.GetBatchMessageCount(localBatchCount - 1)
