@@ -155,7 +155,9 @@ function writeConfigs(argv: any) {
 	const chainInfoFile = path.join(consts.configpath, "l2_chain_info.json")
     const baseConfig = {
         "parent-chain": {
-            "url": argv.l1url,
+            "connection" : {
+                "url": argv.l1url,
+            },
             "wallet": {
                 "account": "",
                 "password": consts.l1passphrase,
@@ -215,9 +217,6 @@ function writeConfigs(argv: any) {
 					"jwtsecret": valJwtSecret,
 				}
             }
-        },
-        "init": {
-            "dev-init": "true"
         },
         "persistent": {
             "chain": "local"
