@@ -5,6 +5,7 @@ package headerreader
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"math/big"
 	"sync"
@@ -19,7 +20,6 @@ import (
 	"github.com/offchainlabs/nitro/arbutil"
 	"github.com/offchainlabs/nitro/solgen/go/precompilesgen"
 	"github.com/offchainlabs/nitro/util/stopwaiter"
-	"github.com/pkg/errors"
 	flag "github.com/spf13/pflag"
 )
 
@@ -311,7 +311,7 @@ func (s *HeaderReader) logIfHeaderIsOld() {
 	l1Timetamp := time.Unix(int64(storedHeader.Time), 0)
 	headerTime := time.Since(l1Timetamp)
 	if headerTime >= s.config().OldHeaderTimeout {
-		s.setError(errors.Errorf("latest header is at least %v old", headerTime))
+		s.setError(fmt.Errorf("latest header is at least %v old", headerTime))
 		log.Warn(
 			"latest L1 block is old", "l1Block", storedHeader.Number,
 			"l1Timestamp", l1Timetamp, "age", headerTime,
